matches_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
asyncpg
python-jose[cryptography]
fastapi-jwt-auth
fastapi-cors
passlib==1.7.4
bcrypt==3.2.2
